test: clear ai_chat/services HIGH review items
- extract realtime transcription test infrastructure to a shared utils
  file (~450 lines)
- WAV-failure test now asserts the real temp-WAV fallback contract
