+++ test/features/ai_chat/services/TEST_REVIEW.md
@@ -24,7 +24,7 @@ Generated by automated test-quality review (static analysis).
 
 - [ ] **[HIGH]** `lib/…/realtime_transcription_service.dart` — 659 lines, the largest implementation file in scope. It currently handles three distinct concerns: Mistral WebSocket session management, MLX local audio channel management, and WAV/M4A file I/O. These three concerns could be split into `mistral_realtime_session.dart`, `mlx_realtime_session.dart`, and a `realtime_audio_writer.dart`, each with its own test file. At 659 lines this is approaching the 1 000-line hard limit and will breach it as MLX support grows.
 
-- [ ] **[HIGH]** `test/…/realtime_transcription_service_test.dart` — 1 626 lines for 43 tests. The `_TestBench`, `_FakeWebSocketChannel`, `_FakeWebSocketSink`, `_FakeMlxAudioChannel`, and `_ControllableRepository` fakes total ~430 lines of infrastructure in a single test file. These could be moved to a `test/features/ai_chat/services/realtime_transcription_test_utils.dart` helper. The test groups (`resolveRealtimeConfig`, `startRealtimeTranscription`, `stop`, `dispose`, `PCM stream error handling`, `WAV file output`, `PCM buffer cap`, `_saveAudio error handling`) are logically independent and could be split into two files: a `realtime_transcription_mistral_test.dart` for Mistral/WebSocket tests and a `realtime_transcription_mlx_test.dart` for MLX tests.
+- [x] **[HIGH]** `test/…/realtime_transcription_service_test.dart` — 1 626 lines for 43 tests. The `_TestBench`, `_FakeWebSocketChannel`, `_FakeWebSocketSink`, `_FakeMlxAudioChannel`, and `_ControllableRepository` fakes total ~430 lines of infrastructure in a single test file. These could be moved to a `test/features/ai_chat/services/realtime_transcription_test_utils.dart` helper. The test groups (`resolveRealtimeConfig`, `startRealtimeTranscription`, `stop`, `dispose`, `PCM stream error handling`, `WAV file output`, `PCM buffer cap`, `_saveAudio error handling`) are logically independent and could be split into two files: a `realtime_transcription_mistral_test.dart` for Mistral/WebSocket tests and a `realtime_transcription_mlx_test.dart` for MLX tests. **RESOLVED (partially):** the ~450 lines of infrastructure (`FakeWebSocketChannel`, `FakeWebSocketSink`, `ControllableRealtimeRepository`, `FakeMlxAudioChannel`, `RealtimeTranscriptionTestBench`, shared constants/helpers) moved to `realtime_transcription_test_utils.dart`, bringing the test file to ~1170 lines. The Mistral/MLX two-file split is rejected: AGENTS.md forbids splitting one source file's tests across multiple files.
 
 - [ ] **[MED]** `test/…/audio_transcription_service_test.dart` — 1 345 lines for 15 tests. The high line count is driven by per-test config setup (provider + model + temp file creation). A shared `_audioTestBench(cloud, chunks)` helper that wires up the standard Gemini config and returns a `(service, dir)` tuple would collapse the 6–8 boilerplate lines into one call per test.
 
@@ -58,7 +58,7 @@ Generated by automated test-quality review (static analysis).
 
 ## Coverage / missing-behavior gaps
 
-- [ ] **[HIGH]** `realtime_transcription_service.dart` WAV-writing failure path (lines ~550–600 in impl) — the test at line 1546 (`returns null audioFilePath when WAV writing fails`) uses an invalid path `/nonexistent/deeply/nested/path/output` and only asserts `result.transcript == 'test'`. It does not assert `result.audioFilePath == null`. The comment says "should gracefully handle the error" but the core assertion is missing.
+- [x] **[HIGH]** `realtime_transcription_service.dart` WAV-writing failure path (lines ~550–600 in impl) — the test at line 1546 (`returns null audioFilePath when WAV writing fails`) uses an invalid path `/nonexistent/deeply/nested/path/output` and only asserts `result.transcript == 'test'`. It does not assert `result.audioFilePath == null`. The comment says "should gracefully handle the error" but the core assertion is missing. **RESOLVED:** the test's premise was half-right — the impl's catch block deliberately keeps the already-written temp WAV instead of losing the recording, so `audioFilePath` is the systemTemp fallback path, not null. The test is renamed to 'falls back to the temp WAV path when the output move fails' and now asserts the non-null `.wav` systemTemp path exists (and cleans it up).
 
 - [ ] **[MED]** `RealtimeTranscriptionService.startRealtimeTranscription` second call while already active — the service has an `isActive` guard, but no test verifies what happens when `startRealtimeTranscription` is called while `isActive == true`. Does it throw, return silently, or replace the existing session?
